@@ -1,82 +1,54 @@
 {
-  "name": "shanoir-ng",
-  "version": "1.0.0",
-  "description": "Shanoir NG",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/fli-iam/shanoir-ng.git"
-  },
+  "name": "shanoir-ng-front",
+  "version": "0.0.0",
   "scripts": {
-    "build": "npm run build:aot && rimraf dist/* && webpack --config config/webpack.config.js --progress --profile --bail",
-    "build:dev": "rimraf dist/* && webpack --config config/webpack.config.js --progress --profile --bail",
-    "build:aot": "rimraf aot && ngc -p tsconfig-aot.json",
-    "i18n": "ng-xi18n --i18nFormat=xlf2",
-    "ngc": "ngc --i18nFile=./messages.xlf --locale=en --i18nFormat=xlf2",
-    "ngc:fr": "ngc --i18nFile=./src/locale/messages.fr.xlf --locale=fr --i18nFormat=xlf2",
-    "start": "webpack-dev-server --config config/webpack.dev.js --progress --profile --watch --content-base src/",
-    "test": "karma start"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "licenses": [
-    {
-      "type": "MIT",
-      "url": "https://github.com/angular/angular.io/blob/master/LICENSE"
-    }
-  ],
+  "private": true,
   "dependencies": {
-    "@angular/animations": "5.1.1",
-    "@angular/cdk": "5.0.1",
-    "@angular/common": "5.1.1",
-    "@angular/compiler": "5.1.1",
-    "@angular/compiler-cli": "5.1.1",
-    "@angular/core": "5.1.1",
-    "@angular/forms": "5.1.1",
-    "@angular/http": "5.1.1",
-    "@angular/material": "5.0.1",
-    "@angular/platform-browser": "5.1.1",
-    "@angular/platform-browser-dynamic": "5.1.1",
-    "@angular/platform-server": "5.1.1",
-    "@angular/router": "5.1.1",
-    "@angular/upgrade": "5.1.1",
-    "@types/node": "8.5.1",
-    "angular-in-memory-web-api": "0.5.2",
-    "angular-sha": "^0.3.3",
-    "core-js": "2.5.3",
-    "js-sha1": "^0.6.0",
-    "mydatepicker": "2.6.1",
-    "reflect-metadata": "0.1.10",
-    "rxjs": "5.5.5",
-    "systemjs": "0.20.19",
-    "zone.js": "0.8.18"
+    "@angular/animations": "~8.2.7",
+    "@angular/cdk": "~8.2.3",
+    "@angular/common": "~8.2.7",
+    "@angular/compiler": "~8.2.7",
+    "@angular/core": "~8.2.7",
+    "@angular/forms": "~8.2.7",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "~8.2.7",
+    "@angular/platform-browser-dynamic": "~8.2.7",
+    "@angular/router": "~8.2.7",
+    "@stomp/ng2-stompjs": "^7.2.0",
+    "hammerjs": "^2.0.8",
+    "mydatepicker": "^2.6.1",
+    "ng-autosize": "^1.1.0",
+    "rxjs": "~6.4.0",
+    "rxjs-compat": "^6.5.3",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^8.3.3",
-    "@ngtools/webpack": "1.9.1",
-    "angular2-autosize": "1.0.2",
-    "angular2-template-loader": "0.6.2",
-    "awesome-typescript-loader": "3.4.1",
-    "compression-webpack-plugin": "1.1.2",
-    "concurrently": "3.5.1",
-    "copy-webpack-plugin": "4.3.0",
-    "css-loader": "0.28.7",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "1.1.6",
-    "html-loader": "0.5.1",
-    "html-webpack-plugin": "2.30.1",
-    "jasmine": "2.8.0",
-    "karma": "1.7.1",
-    "karma-jasmine": "1.1.1",
-    "karma-phantomjs-launcher": "1.0.4",
-    "karma-sourcemap-loader": "0.3.7",
-    "karma-webpack": "2.0.9",
-    "lite-server": "2.3.0",
-    "open-browser-webpack-plugin": "0.0.5",
-    "postcss": "6.0.14",
-    "raw-loader": "0.5.1",
-    "style-loader": "0.19.1",
-    "to-string-loader": "1.1.5",
-    "typescript": "2.4.2",
-    "webpack": "3.10.0",
-    "webpack-dev-server": "2.9.7",
-    "webpack-merge": "4.1.1"
+    "@angular-devkit/build-angular": "~0.803.5",
+    "@angular/cli": "~8.3.5",
+    "@angular/compiler-cli": "~8.2.7",
+    "@angular/language-service": "~8.2.7",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
