@@ -34,7 +34,7 @@ export class BreadcrumbsService {
             private router: Router, 
             private locationStrategy: LocationStrategy,
             private titleService: Title) {
-                
+        
         locationStrategy.onPopState((event: PopStateEvent) => {
             /* detect back & forward browser events and find the target step using its timestamp */
             for (let i=this.steps.length-1; i>=0; i--) {
