@@ -13,12 +13,14 @@
  */
 
 import { Injectable } from '@angular/core';
+import { process } from '../../process';
 
 declare var Keycloak: any;
 
 @Injectable()
 export class KeycloakService {
     static auth: any = {};
+    // static auth: any = { loggedIn: true };
     private gettingToken: boolean = false;
     private tokenPromise: Promise<string>;
 
@@ -28,7 +30,7 @@ export class KeycloakService {
             realm: 'shanoir-ng',
             clientId: 'shanoir-ng-front',
         });
-        KeycloakService.auth.loggedIn = false;
+        KeycloakService.auth.loggedIn = true; // false;
 
         return new Promise((resolve, reject) => {
             keycloakAuth.init({ onLoad: 'login-required' })
