@@ -1,34 +1,24 @@
-/**
- * Shanoir NG - Import, manage and share neuroimaging data
- * Copyright (C) 2009-2019 Inria - https://www.inria.fr/
- * Contact us on https://project.inria.fr/shanoir/
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program. If not, see https://www.gnu.org/licenses/gpl-3.0.html
- */
+import 'hammerjs';
+
+// import 'buffer';
 
 import { enableProdMode } from '@angular/core';
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
+import { environment } from './environments/environment';
 import { KeycloakService } from './app/shared/keycloak/keycloak.service';
 
-// Enable production mode unless running locally
-if (process.env.ENV === 'production') {
+if (environment.production) {
   enableProdMode();
 }
 
 if (window.location.href.endsWith('/account-request')) {
   // Public URL
-  platformBrowserDynamic().bootstrapModule(AppModule);
+  platformBrowserDynamic().bootstrapModule(AppModule).catch(err => console.error(err));
 } else {
   KeycloakService.init()
     .then(() => {
       platformBrowserDynamic().bootstrapModule(AppModule);
-    });
+    }).catch(err => console.error(err));
 }
